# docs/openapi.json
{
  "components": {
    "schemas": {
      "AgentDescriptor": {
        "additionalProperties": false,
        "description": "The catalog entry for one agent, at one version.",
        "properties": {
          "agent_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "approval_threshold": {
            "$ref": "#/components/schemas/ApprovalThreshold",
            "default": "NONE"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "minItems": 1,
            "title": "Capabilities",
            "type": "array",
            "uniqueItems": true
          },
          "classifications_accessed": {
            "items": {
              "$ref": "#/components/schemas/Classification"
            },
            "minItems": 1,
            "title": "Classifications Accessed",
            "type": "array",
            "uniqueItems": true
          },
          "deprecated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deprecated At"
          },
          "input_schema_ref": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Input Schema Ref",
            "type": "string"
          },
          "latency_target_ms": {
            "exclusiveMinimum": 0.0,
            "maximum": 600000.0,
            "title": "Latency Target Ms",
            "type": "integer"
          },
          "loop": {
            "$ref": "#/components/schemas/Loop"
          },
          "output_schema_ref": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Output Schema Ref",
            "type": "string"
          },
          "published_at": {
            "format": "date-time",
            "title": "Published At",
            "type": "string"
          },
          "publisher_department": {
            "$ref": "#/components/schemas/Department"
          },
          "required_scopes": {
            "items": {
              "$ref": "#/components/schemas/Scope"
            },
            "minItems": 1,
            "title": "Required Scopes",
            "type": "array",
            "uniqueItems": true
          },
          "role_summary": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Role Summary",
            "type": "string"
          },
          "version": {
            "maxLength": 32,
            "minLength": 5,
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Version",
            "type": "string"
          },
          "write_targets": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Write Targets",
            "type": "array"
          }
        },
        "required": [
          "agent_id",
          "version",
          "publisher_department",
          "loop",
          "role_summary",
          "capabilities",
          "required_scopes",
          "classifications_accessed",
          "input_schema_ref",
          "output_schema_ref",
          "latency_target_ms",
          "published_at"
        ],
        "title": "AgentDescriptor",
        "type": "object"
      },
      "AgentDescriptorView": {
        "additionalProperties": false,
        "description": "A catalog entry, with every set serialised in sorted order.",
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "approval_threshold": {
            "$ref": "#/components/schemas/ApprovalThreshold"
          },
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "title": "Capabilities",
            "type": "array"
          },
          "classifications_accessed": {
            "items": {
              "type": "string"
            },
            "title": "Classifications Accessed",
            "type": "array"
          },
          "deprecated_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deprecated At"
          },
          "input_schema_ref": {
            "title": "Input Schema Ref",
            "type": "string"
          },
          "latency_target_ms": {
            "title": "Latency Target Ms",
            "type": "integer"
          },
          "loop": {
            "$ref": "#/components/schemas/Loop"
          },
          "output_schema_ref": {
            "title": "Output Schema Ref",
            "type": "string"
          },
          "published_at": {
            "title": "Published At",
            "type": "string"
          },
          "publisher_department": {
            "$ref": "#/components/schemas/Department"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "required_scopes": {
            "items": {
              "type": "string"
            },
            "title": "Required Scopes",
            "type": "array"
          },
          "role_summary": {
            "title": "Role Summary",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "string"
          },
          "write_targets": {
            "items": {
              "type": "string"
            },
            "title": "Write Targets",
            "type": "array"
          }
        },
        "required": [
          "agent_id",
          "version",
          "ref",
          "publisher_department",
          "loop",
          "role_summary",
          "capabilities",
          "required_scopes",
          "classifications_accessed",
          "write_targets",
          "approval_threshold",
          "input_schema_ref",
          "output_schema_ref",
          "latency_target_ms",
          "published_at"
        ],
        "title": "AgentDescriptorView",
        "type": "object"
      },
      "AgentListResponse": {
        "additionalProperties": false,
        "properties": {
          "agents": {
            "items": {
              "$ref": "#/components/schemas/AgentDescriptorView"
            },
            "title": "Agents",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          }
        },
        "required": [
          "agents",
          "count"
        ],
        "title": "AgentListResponse",
        "type": "object"
      },
...
      "ApprovalThreshold": {
        "enum": [
          "NONE",
          "SUPERVISOR",
          "CHIEF"
        ],
        "title": "ApprovalThreshold",
        "type": "string"
      },
...
      "Capability": {
        "enum": [
          "READ",
          "WRITE",
          "RANK",
          "NOTIFY"
        ],
        "title": "Capability",
        "type": "string"
      },
...
      "Classification": {
        "description": "Statutory handling class for a fact.\n\n``PHI`` and ``TIER_II_CONFIDENTIAL`` are the two classes that can never be\nserialized into a vector payload (see :mod:`firstdue.domain.vectors`).",
        "enum": [
          "PUBLIC",
          "RESTRICTED",
          "PHI",
          "TIER_II_CONFIDENTIAL"
        ],
        "title": "Classification",
        "type": "string"
      },
...
      "Department": {
        "description": "Municipal departments that publish agents or receive writes.",
        "enum": [
          "fire",
          "building",
          "planning",
          "county-oem",
          "water",
          "public-works",
          "police",
          "utility",
          "ems"
        ],
        "title": "Department",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Loop": {
        "description": "Which tempo an agent runs at.",
        "enum": [
          "SLOW",
          "INCIDENT"
        ],
        "title": "Loop",
        "type": "string"
      },
...
      "Scope": {
        "description": "Grant scopes. Read scopes and write scopes are deliberately separate:\nan agent authorized to read the permit system is not thereby authorized to\nfile with it.",
        "enum": [
          "read:public-records",
          "read:tier-ii-metadata",
          "read:ems-derived",
          "read:profile",
          "read:geometry",
          "read:audit",
          "write:profile",
          "write:referral",
          "write:work-order",
          "write:preincident-plan",
          "write:rms",
          "notify:agency",
          "write:utility-shutoff",
          "write:road-closure"
        ],
        "title": "Scope",
        "type": "string"
      },
...
      "SubscribeRequest": {
        "additionalProperties": false,
        "description": "Bind a department to one exact version of an agent.",
        "properties": {
          "agent_id": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Agent Id",
            "type": "string"
          },
          "pinned_version": {
            "maxLength": 32,
            "minLength": 5,
            "pattern": "^\\d+\\.\\d+\\.\\d+$",
            "title": "Pinned Version",
            "type": "string"
          },
          "subscriber_department": {
            "$ref": "#/components/schemas/Department"
          },
          "subscription_id": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subscription Id"
          }
        },
        "required": [
          "subscriber_department",
          "agent_id",
          "pinned_version"
        ],
        "title": "SubscribeRequest",
        "type": "object"
      },
      "SubscriptionListResponse": {
        "additionalProperties": false,
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "subscriptions": {
            "items": {
              "$ref": "#/components/schemas/SubscriptionView"
            },
            "title": "Subscriptions",
            "type": "array"
          }
        },
        "required": [
          "subscriptions",
          "count"
        ],
        "title": "SubscriptionListResponse",
        "type": "object"
      },
      "SubscriptionView": {
        "additionalProperties": false,
        "properties": {
          "agent_id": {
            "title": "Agent Id",
            "type": "string"
          },
          "pinned_version": {
            "title": "Pinned Version",
            "type": "string"
          },
          "ref": {
            "title": "Ref",
            "type": "string"
          },
          "subscribed_at": {
            "title": "Subscribed At",
            "type": "string"
          },
          "subscriber_department": {
            "$ref": "#/components/schemas/Department"
          },
          "subscription_id": {
            "title": "Subscription Id",
            "type": "string"
          },
          "unsubscribed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unsubscribed At"
          }
        },
        "required": [
          "subscription_id",
          "subscriber_department",
          "agent_id",
          "pinned_version",
          "ref",
          "subscribed_at"
        ],
        "title": "SubscriptionView",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Municipal structural intelligence as an institutional agent fleet.\n\nA decision-support prototype. It delivers information and performs clerical\nexecution. It never issues tactical recommendations, evacuation orders, crew\nassignments, or fire-behaviour predictions.",
    "title": "FIRST DUE",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/registry/agents": {
      "get": {
        "description": "List the catalog, filtered. Every filter is an ``AND``.",
        "operationId": "list_agents_api_v1_registry_agents_get",
        "parameters": [
          {
            "in": "query",
            "name": "publisher_department",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Department"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Publisher Department"
            }
          },
          {
            "in": "query",
            "name": "loop",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Loop"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Loop"
            }
          },
          {
            "in": "query",
            "name": "capability",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Capability"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Capability"
            }
          },
          {
            "in": "query",
            "name": "agent_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Agent Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Discover agents",
        "tags": [
          "registry"
        ]
      },
      "post": {
        "description": "Publish a descriptor.\n\nIdempotent for an identical descriptor and a 409 for a changed one: a\nversion somebody has pinned must not turn into different code underneath\nthem.",
        "operationId": "publish_agent_api_v1_registry_agents_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AgentDescriptor"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDescriptorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "This version is already published, differently."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Publish an agent descriptor",
        "tags": [
          "registry"
        ]
      }
    },
    "/api/v1/registry/agents/{agent_id}/{version}": {
      "get": {
        "operationId": "get_agent_api_v1_registry_agents__agent_id___version__get",
        "parameters": [
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version",
            "required": true,
            "schema": {
              "title": "Version",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDescriptorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch one published version",
        "tags": [
          "registry"
        ]
      }
    },
    "/api/v1/registry/subscriptions": {
      "get": {
        "operationId": "list_subscriptions_api_v1_registry_subscriptions_get",
        "parameters": [
          {
            "in": "query",
            "name": "subscriber_department",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Department"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Subscriber Department"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List subscriptions",
        "tags": [
          "registry"
        ]
      },
      "post": {
        "description": "Pin a department to one version.\n\nSubscribing to an unpublished version is a 404 rather than a promise to\nresolve later: a pin to something that does not exist is not a pin.",
        "operationId": "subscribe_api_v1_registry_subscriptions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubscriptionView"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "That agent version has not been published."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Subscribe a department to a pinned version",
        "tags": [
          "registry"
        ]
      }
    },
    "/api/v1/registry/subscriptions/{subscriber_department}/{agent_id}/resolved": {
      "get": {
        "description": "The exact descriptor this department runs.\n\nPublishing a newer version does not change this answer. That is the point.",
        "operationId": "resolve_pinned_api_v1_registry_subscriptions__subscriber_department___agent_id__resolved_get",
        "parameters": [
          {
            "in": "path",
            "name": "subscriber_department",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Department"
            }
          },
          {
            "in": "path",
            "name": "agent_id",
            "required": true,
            "schema": {
              "title": "Agent Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentDescriptorView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resolve the pinned version a department runs",
        "tags": [
          "registry"
        ]
      }
    },
...
  }
}